* NewGPUSolver:
  - finally a gpu solver running on vert-frag pipeline is done. However I noticed that when having multiple trees, some tree may become blowed up, while having only one tree does not. Don't know why....
  - move the deprecated shaders out of the project folder for further maybe accidental references.
  Mostly done, have not run yet, definitely have a lot of bugs, need to further debug  - removed shaders that are created while designing the algorithm.

Conflicts:
	Test/GPUTest.unity

+++ Core/Force/Spring2D.cs
@@ -31,7 +31,7 @@ namespace ParticlePhysics2D {
 		/// The convergence group ID, used in gpu solver to group springs which need to be process in parallel
 		/// In cpu solver, this is not needed. 0 means no id is assigned, range 0-255
 		/// </summary>
-		public byte convergenceGroupID = 0;
+		public int convergenceGroupID = 0;
 		
 		// you need to SetSimulation in Simualtion class's OnAfterDeserialize() callback. this is a hack
 		public void SetParticles(Simulation sim) {
